Added makefiles for Mac and Linux
This makefile will handle the executable generation. It should contain
setup and ressources for all the libraries we will use and for the code
we will write. We will detail GLEW and glfw3 later.
For now, it will compile the lodepng helper and our main.cpp
(glm is header-based, and will thus be implicitly compiled when needed)