Added interaction callbacks back
The callbacks are still defined globally, but pass information to the
renderer. This helps to separate the Renderer from the GLFW window.
Moreover, if we wanted to use Renderer methods directly as callbacks
while still having access to the Renderer attributes in those
callbacks, we would need to define a singleton Renderer and so on...


Former-commit-id: 8c8b15a7a17c6f803f3faebe673b8bc8f067b51c