Upload a texture to the GPU
In the initialization, we load the texture from disk, into an array.
Then, a texture 2D storage is used to store this array on the GPU, with
the correct settings (order of colors, size, ...). At draw time, we
active a texture slot, bind the texture, and pass it through a uniform
to the fragment shader. In the fragment shader, a sampler2D uniform is
used, and the texture function allow us to query the color for given UV
coordinates in [0,1]: we use the positionScaled computed earlier for
this.