Update the light position at each frame
The position of the light in view space is now updated in the physics
step. To upload the position to the GPU, one can use glBufferSubData,
but for updates at each frame some hardware can exhibit extremely slow
performance. A more general way is to get a handle to the GPU memory
and update the light position directly.


Former-commit-id: 52fa8a4db85b71b3e644ba9d80f26f5c35ba9e78