Interface blocks
Blocks (structs) can also be used to gather data passed from the vertex
shader to the fragment shader. The user-defined type of the block has
to be the same in both shaders, but the names can be different. This
helps avoiding namespaces collisions: we can now have light.position
and In.position at the same time.


Former-commit-id: dbe7266dc614e11932024684e9f40b8b8b2fb2a9