Normal mapping: GLSL code
In the vertex shader, the matrix to convert positions from tangent
space to view space is computed using the attributes and passed to the
fragment stage. In the fragment shader, the normal is read from the
normal map and converted to view space using the aforementioned matrix.
Shading is same computed in the same way as before.


Former-commit-id: b4ee8491c0a962f538437b94eb18268c11027f97