Cubemap: scale and reverse orientation.
The cube map is scaled to make it bigger, containing the rest of the
scene. As we now are inside the cube, it will be backface culled. Thus,
we reverse the orientation of all triangles by permuting indices to
make the cube visible from the inside. Also adjust the color of the
light.