Uniform buffers: implement double light buffer and remove unused material.
We remove the unused material struct from the C++ code and shaders.
We also make the update of the light position at each frame safer, by
using alternatively two copies of the light struct. This way, we can
perform an unsynchronized write without overwriting data currently
in-use by the GPU (when the previous frame isn't finished for instance).