Shadowmap in the shader
In the shader, we use the light space coordinates of the fragment to
query the corresponding depth reach by light rays in the shadow map,
and we compare it to the depth of the fragment in the same space. If
the fragment is "deeper", it is in the shadow and the diffuse and
specular components of the light should be discarded.


Former-commit-id: 3c3fc35829d8ac992e844ee0a5b7269ce0a4fea9