Edge detection using local luma variations
Query the luma at the four direct neighbors of the current pixel. If
the maximal variation of luma for those pixels (ie the contrast) is
below a given threshold proportional to the maximum local luma, we are
in a uniform area, there is no edge to smooth. Alternatively, if we are
in a dark area, aliasing is less visible and doesn't need to be
smoothed.