Iterate along the light direction.
We ray-march along the light direction until we reach the minimal depth
(0). If, at a step, the current layer depth is bigger than the surface
depth, the sample is under the surface. In this case the shadow amount
is increased based on the difference between the surface and layer
depths, and the current iteration (the samples closer to the inital one
have more weight). Then iterate.