Add a tonemapping pass for future HDR.
After light accumulation and before applying antialiasing, we will
perform a tonemapping pass. Once HDR rendering will be on, it will
allows us to convert HDR values to low-range one at then end. For now,
it simply displays areas where the lighting is > 1 in red. As we are
still in a LDR pipeline, there is none for now.