Add shadows for point lights, using layered framebuffers and geometry shaders.
Point lights can now use a cube shadow map capturing their whole surrounding. To render to such a cubemap framebuffer, we use a geometry shader that project each triangle to the 6 layers of the cubemap, using the proper viewproj matrices. In the shadow map fragment shader, we output the distance from the current point to the light in world space, to simplify future computations. We normalize it using the far plane distance to obtain values in [0,1].


Former-commit-id: 09a1c774eb874eefdf50164823883506c516953e