Raycaster: create a bounding volume hierarchy.
The mesh triangles are binned into a tree hierarchy of bounding volumes (AABB here). The split is done along a random axis at each depth.
At intersection time, ray-box intersection tests are done to go deeper in the BVH, until a leaf containing triangles is reached. Intersection tests are done only for those leaves, and the closest is kept.


Former-commit-id: 8030dd4e378744bc67745fe4ab448b7ca7322172