Engine: handle coordinates flipping in the projection, avoid flipping images at load
This avoids having to flip the y coordinate in every shader, which was causing issues with screen passes. We also stop flipping images at load time, meaning that any texture can be displayed in the right orientation out-of-the-box (whether by blitting or using a screen quad), whether it is loaded, created on the CPU, or coming from a framebuffer.