PBRDemo: add anisotropic material.
This material exhibits a non-isotropic behavior when reflecting light.

Anisotropy is parametrized by a direction (rotation of the local tangent frame) and an intensity, parametrized as described by C. Kulla and A. Conty in "Revisiting Physically Based Shading at Imageworks", 2017 (http://www.aconty.com/pdf/s2017_pbs_imageworks_slides.pdf).

A specific anisotropic version of the GGX BRDF is used, and a bent reflection hack is added to simulate anisotropic reflection probes as described by S. McAuley in "Rendering the World of Far Cry 4", 2015 (https://www.gdcvault.com/play/1022235/Rendering-the-World-of-Far).