PBRDemo: add material with subsurface scattering.
We add an extra lobe towards the other side of the object, with its own roughness and transmittance.
This is inspired from the technique described by C. Barré-Brisebois in "Approximating Translucency for a Fast, Cheap and Convincing Subsurface-Scattering Look", 2011, (https://colinbarrebrisebois.com/2011/03/07/gdc-2011-approximating-translucency-for-a-fast-cheap-and-convincing-subsurface-scattering-look/), but using a GGX distribution.
This also motivates the use of SH irradiance to estimate the transmitted radiance when evaluating the ambient contribution of subsurface scattering.